requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core==1.14.0
azure-functions==1.7.0
azure-search-documents==11.1.0
certifi==2021.5.30
chardet==4.0.0
idna==2.10
isodate==0.6.0
msrest==0.6.21
numpy==1.20.3
oauthlib==3.1.1
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.5
